* Linked list copy semantics removed, DoubleLink added * BYTE is not uppercase anymore * Reduced the number of warnings * Copyright notices to some of the files (more coming later) * EventChain now has a clear namespace blueprint * LinkedList now has item iterator along with object iterator * LinkedList items and iterators now have STL like (*, ->) and regular type conversions * ConsumableEvent coding part I

+++ Resource/ResizableObject.h
@@ -49,11 +49,11 @@ namespace gre {
 				if(Type==Integral) {
 					switch(CalculationMethod) {
 						case SmallerThanGivenSize:
-							return gge::Max(floor((float)Requested/Original), 1)*Original;
+							return (int)gge::Max(floor((float)Requested/Original), 1)*Original;
 						case ClosestToGivenSize:
-							return gge::Max(gge::Round((float)Requested/Original), 1)*Original;
+							return (int)gge::Max(gge::Round((float)Requested/Original), 1)*Original;
 						case FillTheGivenSize:
-							return gge::Max(ceil((float)Requested/Original), 1)*Original;
+							return (int)gge::Max(ceil((float)Requested/Original), 1)*Original;
 					}
 				}
 
@@ -94,11 +94,11 @@ namespace gre {
 				}
 
 				if(CalculationMethod==SmallerThanGivenSize)
-					return gge::Max(floor( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
+					return (int)gge::Max(floor( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
 				else if(CalculationMethod==ClosestToGivenSize)
-					return gge::Max(gge::Round( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
+					return (int)gge::Max(gge::Round( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
 				else
-					return gge::Max(ceil( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
+					return (int)gge::Max(ceil( ((float)(w-Overhead)) / Increment ), 0)  * Increment + Overhead;
 			}
 		};
